split out module for supported kits, added expert flag
