+++ rna_extraction_kits/libraries/supportedrnaextractionkits/definition.json
@@ -1,0 +1,7 @@
+{
+  "id": 43,
+  "name": "SupportedRNAExtractionKits",
+  "parent_class": "Library",
+  "category": "RNA Extraction Kits",
+  "user_id": 1
+}
